# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/stores": {
      "get": {
        "operationId": "StoreController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "페이지 번호",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "페이지당 항목 수",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "name",
            "required": false,
            "in": "query",
            "description": "상점 이름 검색",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "description": "카테고리 ID로 필터링",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "ownerId",
            "required": false,
            "in": "query",
            "description": "점주 ID로 필터링",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "활성 상태로 필터링",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "isVerified",
            "required": false,
            "in": "query",
            "description": "인증 상태로 필터링",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "정렬 기준",
            "schema": {
              "type": "string",
              "enum": [
                "createdAt",
                "name",
                "categoryId"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "정렬 방향",
            "schema": {
              "type": "string",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "상점 목록 반환"
          }
        },
        "summary": "상점 목록 조회",
        "tags": [
          "상점"
        ]
      },
      "post": {
        "operationId": "StoreController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": "상점 등록 성공"
          },
          "400": {
            "description": "잘못된 요청"
          },
          "401": {
            "description": "인증되지 않은 요청"
          },
          "403": {
            "description": "권한 없음"
          },
          "409": {
            "description": "사업자 등록 번호 중복"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "상점 등록",
        "tags": [
          "상점"
        ]
      }
    },
...
    "/api/stores/{storeId}/operating-hours": {
      "get": {
        "operationId": "OperatingHoursController_findAll",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "description": "매장 ID",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "영업시간 목록 반환",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperatingHoursResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "매장을 찾을 수 없음"
          }
        },
        "summary": "매장 영업시간 목록 조회",
        "tags": [
          "매장 영업시간"
        ]
      },
      "post": {
        "operationId": "OperatingHoursController_create",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOperatingHourDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "영업시간 등록 성공"
          },
          "400": {
            "description": "잘못된 요청"
          },
          "401": {
            "description": "인증되지 않은 요청"
          },
          "403": {
            "description": "권한 없음"
          },
          "404": {
            "description": "매장을 찾을 수 없음"
          },
          "409": {
            "description": "이미 존재하는 요일 영업시간"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "매장 영업시간 등록 (개별)",
        "tags": [
          "매장 영업시간"
        ]
      },
      "delete": {
        "operationId": "OperatingHoursController_removeAll",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "영업시간 모두 삭제 성공"
          },
          "401": {
            "description": "인증되지 않은 요청"
          },
          "403": {
            "description": "권한 없음"
          },
          "404": {
            "description": "매장을 찾을 수 없음"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "매장 영업시간 모두 삭제",
        "tags": [
          "매장 영업시간"
        ]
      }
    },
    "/api/stores/{storeId}/operating-hours/{day}": {
      "get": {
        "operationId": "OperatingHoursController_findByDay",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "description": "매장 ID",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "day",
            "required": true,
            "in": "path",
            "description": "요일",
            "schema": {
              "enum": [
                "monday",
                "tuesday",
                "wednesday",
                "thursday",
                "friday",
                "saturday",
                "sunday"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "영업시간 정보 반환"
          },
          "404": {
            "description": "매장 또는 영업시간을 찾을 수 없음"
          }
        },
        "summary": "매장 특정 요일 영업시간 조회",
        "tags": [
          "매장 영업시간"
        ]
      }
    },
    "/api/stores/{storeId}/operating-hours/bulk": {
      "post": {
        "operationId": "OperatingHoursController_createBulk",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "영업시간 일괄 등록 성공"
          },
          "400": {
            "description": "잘못된 요청"
          },
          "401": {
            "description": "인증되지 않은 요청"
          },
          "403": {
            "description": "권한 없음"
          },
          "404": {
            "description": "매장을 찾을 수 없음"
          },
          "409": {
            "description": "이미 존재하는 요일 영업시간"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "매장 영업시간 일괄 등록",
        "tags": [
          "매장 영업시간"
        ]
      }
    },
    "/api/stores/{storeId}/operating-hours/{id}": {
      "patch": {
        "operationId": "OperatingHoursController_update",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOperatingHourDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "영업시간 수정 성공"
          },
          "400": {
            "description": "잘못된 요청"
          },
          "401": {
            "description": "인증되지 않은 요청"
          },
          "403": {
            "description": "권한 없음"
          },
          "404": {
            "description": "영업시간을 찾을 수 없음"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "매장 영업시간 수정",
        "tags": [
          "매장 영업시간"
        ]
      },
      "delete": {
        "operationId": "OperatingHoursController_remove",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "영업시간 삭제 성공"
          },
          "401": {
            "description": "인증되지 않은 요청"
          },
          "403": {
            "description": "권한 없음"
          },
          "404": {
            "description": "영업시간을 찾을 수 없음"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "매장 영업시간 삭제",
        "tags": [
          "매장 영업시간"
        ]
      }
    },
    "/api/stores/{storeId}/operating-hours/day/{day}": {
      "patch": {
        "operationId": "OperatingHoursController_updateByDay",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "day",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOperatingHourDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "영업시간 수정 성공"
          },
          "400": {
            "description": "잘못된 요청"
          },
          "401": {
            "description": "인증되지 않은 요청"
          },
          "403": {
            "description": "권한 없음"
          },
          "404": {
            "description": "영업시간을 찾을 수 없음"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "매장 특정 요일 영업시간 수정",
        "tags": [
          "매장 영업시간"
        ]
      },
      "delete": {
        "operationId": "OperatingHoursController_removeByDay",
        "parameters": [
          {
            "name": "storeId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "day",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "영업시간 삭제 성공"
          },
          "401": {
            "description": "인증되지 않은 요청"
          },
          "403": {
            "description": "권한 없음"
          },
          "404": {
            "description": "영업시간을 찾을 수 없음"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "매장 특정 요일 영업시간 삭제",
        "tags": [
          "매장 영업시간"
        ]
      }
    },
...
  },
  "info": {
    "title": "픽업 서버 API",
    "description": "픽업 서버 API 문서",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "OperatingHoursResponseDto": {
        "type": "object",
        "properties": {
          "operatingHours": {
            "description": "영업시간 정보 배열",
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "operatingHoursByDay": {
            "type": "object",
            "description": "영업시간 정보 (요일별로 그룹화)"
          }
        },
        "required": [
          "operatingHours",
          "operatingHoursByDay"
        ]
      },
      "CreateOperatingHourDto": {
        "type": "object",
        "properties": {
          "dayOfWeek": {
            "type": "string",
            "enum": [
              "monday",
              "tuesday",
              "wednesday",
              "thursday",
              "friday",
              "saturday",
              "sunday"
            ],
            "example": "monday",
            "description": "요일"
          },
          "openingTime": {
            "type": "string",
            "example": "09:00",
            "description": "오픈 시간 (24시간 형식, HH:MM)"
          },
          "closingTime": {
            "type": "string",
            "example": "18:00",
            "description": "마감 시간 (24시간 형식, HH:MM)"
          },
          "breakStartTime": {
            "type": "string",
            "example": "12:00",
            "description": "휴게 시작 시간 (24시간 형식, HH:MM)"
          },
          "breakEndTime": {
            "type": "string",
            "example": "13:00",
            "description": "휴게 종료 시간 (24시간 형식, HH:MM)"
          },
          "isDayOff": {
            "type": "boolean",
            "example": false,
            "description": "휴무일 여부"
          }
        },
        "required": [
          "dayOfWeek",
          "openingTime",
          "closingTime"
        ]
      },
      "UpdateOperatingHourDto": {
        "type": "object",
        "properties": {
          "openingTime": {
            "type": "string",
            "example": "09:00",
            "description": "오픈 시간 (24시간 형식, HH:MM)"
          },
          "closingTime": {
            "type": "string",
            "example": "18:00",
            "description": "마감 시간 (24시간 형식, HH:MM)"
          },
          "breakStartTime": {
            "type": "string",
            "example": "12:00",
            "description": "휴게 시작 시간 (24시간 형식, HH:MM)"
          },
          "breakEndTime": {
            "type": "string",
            "example": "13:00",
            "description": "휴게 종료 시간 (24시간 형식, HH:MM)"
          },
          "isDayOff": {
            "type": "boolean",
            "example": false,
            "description": "휴무일 여부"
          }
        }
      },
...
    }
  }
}